+++ tests/test_pyproject.py
@@ -4,4 +4,4 @@ import pyproject
 
 def test_version():
     """Verify we have updated the package version."""
-    assert pyproject.__version__ == "21.1.1"
+    assert pyproject.__version__ == "2021.2.0"