+++ CHANGELOG.md
@@ -6,10 +6,15 @@ Versions follow [CalVer](https://calver.org).
 
 ### Added
 
+- Add a GitHub action to run the tests on every push to the main branch.
+- Enable support to run code quality checks using the `pre-commit` library.
 - Add development tasks to install, uninstall, and upgrade the project package.
 
 ### Changed
 
+- In the package entry file, `src/pyproject/__init__.py`, the `__name__`
+  variable must match the name set in the `tool.poetry` section of the
+  `pyproject.toml` file.
 - Use `click` to create the command line interfaces for development tasks.
 
 ### Deprecated