+++ CHANGELOG.md
@@ -6,12 +6,15 @@ Versions follow [CalVer](https://calver.org).
 
 ### Added
 
+- Add `rope` library for development purposes.
+- Add documentation source code using `sphinx` and `sphinx_rtd_theme`.
 - Add a GitHub action to run the tests on every push to the main branch.
 - Enable support to run code quality checks using the `pre-commit` library.
 - Add development tasks to install, uninstall, and upgrade the project package.
 
 ### Changed
 
+- Make the package metadata a public name of the project package.
 - In the package entry file, `src/pyproject/__init__.py`, the `__name__`
   variable must match the name set in the `tool.poetry` section of the
   `pyproject.toml` file.
@@ -27,7 +30,9 @@ TODO.
 
 ### Fixed
 
-TODO.
+- Fix tasks to install, uninstall, and upgrade the project package.
+
+---
 
 ## 21.1.0 (2021-07-24)